refactor(runtime): split host input regression suites
