+++ src/runtime/src/runtime/test_support/values.rs
@@ -1,4 +1,6 @@
-use mech_core::{Value, hash_str};
+use mech_core::{
+    ReactiveCellId, ReactiveDependencyKind, ReactiveNodeId, ReactiveNodeKind, Value, hash_str,
+};
 
 use super::super::MechRuntime;
 use crate::{HostArgumentValue, RuntimeHostInputSource};
@@ -68,4 +70,71 @@ pub(crate) fn source_value(runtime: &MechRuntime, source: &RuntimeHostInputSourc
         .unwrap_or_else(|| panic!("missing symbol {}", input.symbol_id))
         .borrow()
         .clone()
+}
+
+pub(crate) fn symbol_cell(runtime: &MechRuntime, name: &str) -> ReactiveCellId {
+    let cells = symbol_value(runtime, name).reactive_root_cell_ids();
+    assert_eq!(cells.len(), 1, "symbol {name} must have one root cell");
+    cells[0]
+}
+
+pub(crate) fn source_cell(
+    runtime: &MechRuntime,
+    source: &RuntimeHostInputSource,
+) -> ReactiveCellId {
+    let cells = source_value(runtime, source).reactive_root_cell_ids();
+    assert_eq!(cells.len(), 1, "source must have one root cell");
+    cells[0]
+}
+
+pub(crate) fn register_node_for_symbol(runtime: &MechRuntime, name: &str) -> ReactiveNodeId {
+    let output = symbol_cell(runtime, name);
+    let plan = runtime.program.interpreter().plan();
+    let plan = plan.borrow();
+    let nodes = plan
+        .nodes
+        .iter()
+        .filter(|node| node.kind == ReactiveNodeKind::Register && node.outputs.contains(&output))
+        .map(|node| node.id)
+        .collect::<Vec<_>>();
+    assert_eq!(nodes.len(), 1, "symbol {name} must have one register node");
+    nodes[0]
+}
+
+pub(crate) fn combinational_node_for_output_and_inputs(
+    runtime: &MechRuntime,
+    output: ReactiveCellId,
+    required_inputs: &[ReactiveCellId],
+) -> ReactiveNodeId {
+    let plan = runtime.program.interpreter().plan();
+    let plan = plan.borrow();
+    let nodes = plan
+        .nodes
+        .iter()
+        .filter(|node| {
+            node.kind == ReactiveNodeKind::Combinational
+                && node.outputs.contains(&output)
+                && required_inputs.iter().all(|required| {
+                    node.inputs.iter().any(|dependency| {
+                        dependency.cell == *required
+                            && dependency.kind == ReactiveDependencyKind::Reactive
+                    })
+                })
+        })
+        .map(|node| node.id)
+        .collect::<Vec<_>>();
+    assert_eq!(nodes.len(), 1, "expected one matching combinational node");
+    nodes[0]
+}
+
+pub(crate) fn plan_snapshot(
+    runtime: &MechRuntime,
+) -> (usize, Vec<ReactiveNodeId>, Vec<Vec<ReactiveCellId>>) {
+    let plan = runtime.program.interpreter().plan();
+    let plan = plan.borrow();
+    (
+        plan.len(),
+        plan.nodes.iter().map(|node| node.id).collect(),
+        plan.nodes.iter().map(|node| node.outputs.clone()).collect(),
+    )
 }
